Made wrappers check the order is correct (higher courses wrap lower courses)

+++ commands/delwrapper.js
@@ -7,14 +7,18 @@ module.exports = {
   description: "Removes a wrapper course.",
   usage: "`!delwrapper <course>`",
   async execute(message, args) {
+    const filePath = __dirname + "/../data/wrappers.json";
     args[0] = args[0].toLowerCase();
+    if (fs.existsSync(filePath)) {
+      const wrappers = require(filePath);
+      delete wrappers[args[0]];
+      fs.writeFileSync(
+        filePath,
+        JSON.stringify(wrappers)
+      );
+      return message.channel.send(`Removed ${args[0]} as a wrapper.`);
+    }
+    return message.channel.send(`No wrappers found!`);
 
-    const wrappers = require("../data/wrappers.json");
-    delete wrappers[args[0]];
-    fs.writeFileSync(
-      __dirname + "/../data/wrappers.json",
-      JSON.stringify(wrappers)
-    );
-    return message.channel.send(`Removed ${args[0]} as a wrapper.`);
   }
 };
